% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cov_data}
\alias{cov_data}
\title{Coverage Data}
\format{
\subsection{\code{cov_data}}{

A data frame with 906 rows and 111 columns:
}
}
\usage{
cov_data
}
\description{
Coverage Data
}
\keyword{datasets}
