% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ele_cov}
\alias{ele_cov}
\title{Coverage profile of elements}
\format{
\subsection{\code{ele_cov}}{

A data frame with 11 rows and 15 columns:
\describe{
\item{Chr}{Chromosome}
\item{Start}{Start position}
\item{End}{End position}
\item{Length}{Length}
\item{Annotation}{Annotation of elements}
\item{sampleN}{Coverage}
}
}
}
\usage{
ele_cov
}
\description{
Coverage profile of elements
}
\keyword{datasets}
