% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ele_depth}
\alias{ele_depth}
\title{Depth data}
\format{
\subsection{\code{ele_depth}}{

A data frame with 3034 rows and 12 columns
}
}
\usage{
ele_depth
}
\description{
Depth data
}
\keyword{datasets}
