% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ele_pav}
\alias{ele_pav}
\title{PAV profile of elements}
\format{
\subsection{\code{ele_pav}}{

A data frame with 11 rows and 15 columns:
\describe{
\item{Chr}{Chromosome}
\item{Start}{Start position}
\item{End}{End position}
\item{Length}{Length}
\item{Annotation}{Annotation of elements}
\item{sampleN}{PAV}
}
}
}
\usage{
ele_pav
}
\description{
PAV profile of elements
}
\keyword{datasets}
