% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ele_pheno}
\alias{ele_pheno}
\title{Phenotype data}
\format{
\subsection{\code{ele_pheno}}{

A data frame with 10 rows and 3 columns
}
}
\usage{
ele_pheno
}
\description{
Phenotype data
}
\keyword{datasets}
