% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{est_group_res}
\alias{est_group_res}
\title{Estimation result with grouped samples}
\format{
\subsection{\code{est_group_res}}{

A data frame with 1110 rows and 6 columns:
\describe{
\item{Round}{Estimation round}
\item{SampleN}{The number of selected samples}
\item{Core}{The number of core regions}
\item{Pan}{The number of all regions}
\item{Delta}{The number of novel regions}
\item{Group}{The group of samples}
}
}
}
\usage{
est_group_res
}
\description{
Estimation result with grouped samples
}
\keyword{datasets}
