% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{est_res}
\alias{est_res}
\title{Estimation result}
\format{
\subsection{\code{est_res}}{

A data frame with 1110 rows and 5 columns:
\describe{
\item{Round}{Estimation round}
\item{SampleN}{The number of selected samples}
\item{Core}{The number of core regions}
\item{Pan}{The number of all regions}
\item{Delta}{The number of novel regions}
}
}
}
\usage{
est_res
}
\description{
Estimation result
}
\keyword{datasets}
