% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_info_data}
\alias{gene_info_data}
\title{Gene information}
\format{
\subsection{\code{gene_info_data}}{

A data frame with 906 rows and 5 columns:
\describe{
\item{Chr}{Chromosome}
\item{start}{Start position}
\item{end}{End position}
\item{length}{Length}
\item{chr_n}{The Number of Chromosome}
}
}
}
\usage{
gene_info_data
}
\description{
Gene information
}
\keyword{datasets}
