% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_obj.R
\name{get_cov_obj}
\alias{get_cov_obj}
\title{Get an object of the COV class.}
\usage{
get_cov_obj(cov_data, pheno_info = NULL, region_info = NULL)
}
\arguments{
\item{cov_data}{A numeric \code{matrix} or \code{data.frame} of the coverage data.
The row names are region names. The column names are sample names.}

\item{pheno_info}{A \code{data.frame} of phenotype data. The row names are sample name.}

\item{region_info}{A \code{data.frame} of region data. The row names are region name.

The \code{cov_data} shouldn't contain region that is 0 in all sample, and the same for sample.}
}
\description{
Get an object of the COV class.
}
