% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pav_obj.R
\name{get_pav_obj}
\alias{get_pav_obj}
\title{Get an object of the PAV class.}
\usage{
get_pav_obj(
  pav_data,
  pheno_info = NULL,
  region_info = NULL,
  add_softcore = T,
  add_private = T,
  softcore_loss_rate = 0.1,
  use_binomial = F,
  softcore_p_value = 0.05
)
}
\arguments{
\item{pav_data}{A numeric \code{matrix} or \code{data.frame} of PAV data. \code{0} represent "absence" and \code{1} represent "presence".
The row names are region names. The column names are sample names.}

\item{pheno_info}{The phenotype data. The row names are sample names.}

\item{region_info}{Relevant information on target regions. The row names are region names.}

\item{add_softcore}{A logical value indicating whether to consider \code{soft-core} when determining the category of the region.}

\item{add_private}{A logical value indicating whether to consider \code{private} when determining category of the region.}

\item{softcore_loss_rate}{A numeric vector of loss rate.}

\item{use_binomial}{A logical value indicating whether to use binomial test.}

\item{softcore_p_value}{A numeric vector of p-value (binomial test). It only works when \code{use_binomial}=T

\code{add_softcore} and \code{add_private} are logical values indicating whether to consider "softcore" and "private" when determining the category.

If \code{add_softcore} is \code{TRUE}, the regions with loss rates not significantly larger than \code{softcore_loss_rate} will be considered as "softcore". Binomial tests (with a null hypothesis of loss rate < \code{softcore_loss_rate}) are carried out for each region if \code{use_binomial} is \code{T}. A p-vlaue below \code{softcore_p_value} means that this region is lost in a significant proportion and is a distributed region(loss rate > \code{softcore_loss_rate}).

If \code{add_private} is \code{TRUE}, the regions present in only one sample will be considered as "private" region.}
}
\description{
Get an object of the PAV class.
}
