% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pav_data}
\alias{pav_data}
\title{PAV Data}
\format{
\subsection{\code{pav_data}}{

A data frame with 906 rows and 111 columns:
}
}
\usage{
pav_data
}
\description{
"1" represents presence and "0" represents absence.
}
\keyword{datasets}
