% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno_info_data}
\alias{pheno_info_data}
\title{Phenotype Data}
\format{
\subsection{\code{pheno_info_data}}{

A data frame with 111 rows and 5 columns:
\describe{
\item{Genetic_sex}{Genetic sex}
\item{DNA_source}{DNA source}
\item{Region}{Region}
\item{Country}{Country}
\item{Coverage_mean}{Average of coverage}
}
}
}
\source{
\url{https://www.nature.com/articles/nature18964}
}
\usage{
pheno_info_data
}
\description{
Phenotype Data
}
\keyword{datasets}
