% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_plot.R
\name{pheno_manhattan}
\alias{pheno_manhattan}
\title{Draw a Manhattan plot}
\usage{
pheno_manhattan(
  pav_obj,
  pheno_stat_res,
  pheno_name,
  chr,
  bp,
  adjust_p = T,
  highlight_top_n = 5,
  highlight_text_size = 4,
  point_size = 1.5,
  x_text_size = NULL,
  x_text_angle = 0,
  x_title_size = NULL,
  y_text_size = NULL,
  y_title_size = NULL
)
}
\arguments{
\item{pav_obj}{A PAV object.}

\item{pheno_stat_res}{The result from \code{\link[APAVplot]{pheno_stat}}.}

\item{pheno_name}{The name of phenotype used for displaying.}

\item{chr}{The name in \code{region_info} denoting chromosomes.}

\item{bp}{The name in \code{region_info} denoting chromosomal position.}

\item{adjust_p}{A logical value indicating whether adjust p_value.}

\item{highlight_top_n}{The top \code{n} points will be highlighted.}

\item{highlight_text_size}{The size of labels on highlight points.}

\item{point_size}{The size of points.}

\item{x_text_size}{The size of tick labels on x-axis.}

\item{x_text_angle}{The angle of tick labels.}

\item{x_title_size}{The size of x-axis title.}

\item{y_text_size}{The size of tick labels on y-axis.}

\item{y_title_size}{The size of y-axis title.}
}
\description{
Show the result of phenotype association of a certain phenotype in a Manhattan plot.
}
