% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ele_plot.R
\name{plot_ele_depth}
\alias{plot_ele_depth}
\title{Show element depth}
\usage{
plot_ele_depth(
  depth_data,
  ele_data,
  gff_data = NULL,
  pheno_data = NULL,
  cluster_samples = F,
  clustering_distance = "euclidean",
  clustering_method = "complete",
  log10 = F,
  ele_region = NULL,
  top_anno_height = 0.4,
  left_anno_width = 0.1,
  ele_color = "#A6CEE3",
  depth_colors = c("#5680ae", "white"),
  lowlight_colors = c("gray", "white"),
  gene_colors = structure(c("gray70", "gray85", "#d77478", "#ecbf40", "#ecbf40"), names =
    c("transcript", "exon", "CDS", "five_prime_UTR", "three_prime_UTR")),
  pheno_info_color_list = NULL,
  pheno_border = NA,
  seq_name_size = NULL,
  loc_name_size = NULL,
  show_sample_name = T,
  sample_name_size = NULL,
  legend_title = "Depth",
  legend_title_size = NULL,
  legend_text_size = NULL
)
}
\arguments{
\item{depth_data}{Depth data generated by the APAV tool.}

\item{ele_data}{Element data generated by the APAV tool.}

\item{gff_data}{GFF data.}

\item{pheno_data}{Phenotype data. The row names are sample names.}

\item{cluster_samples}{A logical value indicating whether perform clustering on samples.}

\item{clustering_distance}{Method of measuring distance when clustering, pass to \code{\link[stats]{dist}}.}

\item{clustering_method}{Method to perform hierarchical clustering, pass to \code{\link[stats]{hclust}}.}

\item{log10}{A logical value indicating whether convert depth value to log10(depth+1).}

\item{ele_region}{Gene element regions for display, choosing from "TN"(transcripts number), "UP"(upstream) and "DOWN"(downstream).}

\item{top_anno_height}{The relative height of annotation, ranging from 0 to 1.}

\item{left_anno_width}{The relative width of annotation, ranging from 0 to 1.}

\item{ele_color}{The color of element regions.}

\item{depth_colors}{A vector of colors of depth.}

\item{lowlight_colors}{A vector of colors for depth of non-element region.
Element region not highlighted when set to NULL.}

\item{gene_colors}{A named vector for gene elements.
e.g. c(CDS = "#d77478", five_prime_UTR = "#ecbf40")}

\item{pheno_info_color_list}{A list contains named vector of colors for phenotype.
e.g. list(gender = c("Male" = "green", "Female" = "red"), age = c("yellow", "red"))}

\item{pheno_border}{The color of cell borders in phenotype annotation.}

\item{seq_name_size}{The size of sequence name.}

\item{loc_name_size}{The size of coordinate numbers.}

\item{show_sample_name}{A logical value indicating whether show sample names.}

\item{sample_name_size}{The size of sample names.}

\item{legend_title}{The title of legend.}

\item{legend_title_size}{The size of legend title.}

\item{legend_text_size}{The size of legend item labels.}
}
\description{
Plot the depth of elements
}
