% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ele_plot.R
\name{plot_ele_pav}
\alias{plot_ele_pav}
\title{Show element PAV}
\usage{
plot_ele_pav(
  data,
  gff_data = NULL,
  pheno_data = NULL,
  cluster_samples = F,
  clustering_distance = "euclidean",
  clustering_method = "complete",
  ele_region = NULL,
  top_anno_height = 0.4,
  left_anno_width = 0.1,
  ele_color = "#A6CEE3",
  ele_line_color = "gray",
  pav_colors = c(presence = "#5680ae", absence = "gray70"),
  cell_border = NA,
  gene_colors = structure(c("gray70", "gray85", "#d77478", "#ecbf40", "#ecbf40"), names =
    c("transcript", "exon", "CDS", "five_prime_UTR", "three_prime_UTR")),
  pheno_info_color_list = NULL,
  pheno_border = NA,
  seq_name_size = NULL,
  loc_name_size = NULL,
  show_ele_name = T,
  ele_name_size = NULL,
  ele_name_rot = 90,
  ele_name_len = 20,
  show_sample_name = T,
  sample_name_size = NULL,
  legend_title_size = NULL,
  legend_text_size = NULL
)
}
\arguments{
\item{data}{Coverage results data generated by the APAV tool.}

\item{gff_data}{GFF data.}

\item{pheno_data}{Phenotype data. The row names are sample names.}

\item{cluster_samples}{A logical value indicating whether perform clustering on samples.}

\item{clustering_distance}{Method of measuring distance when clustering, pass to \code{\link[stats]{dist}}.}

\item{clustering_method}{Method to perform hierarchical clustering, pass to \code{\link[stats]{hclust}}.}

\item{ele_region}{Gene element regions for display, choosing from "TN"(transcripts number), "UP"(upstream) and "DOWN"(downstream).}

\item{top_anno_height}{The relative height of annotation, ranging from 0 to 1.}

\item{left_anno_width}{The relative width of annotation, ranging from 0 to 1.}

\item{ele_color}{The color of element regions.}

\item{ele_line_color}{The color for the line pointing to elements.}

\item{pav_colors}{A named vector of colors for presence and absence.
e.g. c(presence = "#5680ae", absence = "gray70")}

\item{cell_border}{The color of cell borders in heat map.}

\item{gene_colors}{A named vector for gene elements.
e.g. c(CDS = "#d77478", five_prime_UTR = "#ecbf40")}

\item{pheno_info_color_list}{A list contains named vector of colors for phenotype.
e.g. list(gender = c("Male" = "green", "Female" = "red"), age = c("yellow", "red"))}

\item{pheno_border}{The color of cell borders in phenotype annotation.}

\item{seq_name_size}{The size of sequence name.}

\item{loc_name_size}{The size of coordinate numbers.}

\item{show_ele_name}{A logical value indicating whether show element names.}

\item{ele_name_size}{The size of element names.}

\item{ele_name_rot}{The rotation of element names.}

\item{ele_name_len}{The longest length of element names.}

\item{show_sample_name}{A logical value indicating whether show sample names.}

\item{sample_name_size}{The size of sample names.}

\item{legend_title_size}{The size of legend title.}

\item{legend_text_size}{The size of legend item labels.}
}
\description{
Plot the PAV of elements.
}
