% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_res}
\alias{sim_res}
\title{Simulation result}
\format{
\subsection{\code{sim_res}}{

A data frame with 1110 rows and 5 columns:
\describe{
\item{Round}{Simulation round}
\item{SampleN}{The number of selected samples}
\item{Core}{The number of core regions}
\item{Pan}{The number of all regions}
\item{Delta}{The number of novel regions}
}
}
}
\usage{
sim_res
}
\description{
Simulation result
}
\keyword{datasets}
